Work around type promotion issues with julia 0.4 in gensolve().
Julia 0.4 seems to bahve a little differently when sparse matrices with
different element types are involved in a computation than julia 0.3.
For example, dividing a SparseMatrixCSC{Number,Int64} by a Float64 now
gives a SparseMatrixCSC{Any,Int64} instead of a
SparseMatrixCSC{Number,Int64}. Multiplication, however, has not changed,
so A/x is replaced with A*(1/x). Furthoermore, explicit calls to convert
are used to ansure that x and h keep their type in gesolve().