Improvements to k-d tree.
The number of nodes in the tree is no longer padded to the next power of
two (by replicating part of the input data). Instead, the lowest level
of the tree is filled from left to right.

Additionally, the boundaries are no longer set on the data values, but
in the middle between them.