CachingSolver: Improve heuristics when to re-create the tree.
The number of solutions stored in the array for linear search is now
limited proportional to the number of entries in the tree. Furthermore,
the number is now reduced for every invocation of solve, not only when a
new solution is cached, to be sure to eventually get all solutions into
the tree.