Cope with underdetermined linear part
For pathological circuits, the linear equation part may be
underdetermined. E.g. for a shorted 0Ω resistor, the current is
inderterminate. Instead of growing the z vector, which makes the
non-linear solver fail, the inderterminate quantitity is now ignored
(assumened to be zero) and a warning is printed if outout or state update
depend on it.