ParametricNonLinEq: Seperate setting of p and evaluation
ParametricNonLinEq provides a scratch area that can be used when p is set
and when the function is evaluated to forward the information of p.
Currently, the default implementation of copying is used throughout.

When solving, p is not set within the inner loop, but only once at the
beginning.