Eliminate constant non-linearities
If a non-linearity does not depend on u, x, or previous z's, its
solution will always be the initial solution. So can be entirely
replaced with an updated q0. This may allow other non-linearites to be
identified as constant if they only depended on the solution to this, so
the process is iterated.