Use Gummel-Poon model for BJTs
The Gummel-Poon model is a strict generalization of the previously used
Ebers-Moll model and the default values for the new parameters are
chosen such that they result in the Ebers-Moll case. However, the
definition of the saturation currents is slightly different than what is
usually found in the Gummel-Poon model (scaled by a factor of β/(1+β)).