Fix matrix splitting for decomposed nonlinearity
The algorithm assumed the rows corresponding to the sub-nonlinearites to
be ordered from top to bottom, but the ordering can actually be
arbitrary and has to be derived from the decomposition. However, for
many circuits, the dependencies match the natural order by which the
circuit is entered, and then the decomposition order actually becomes
top to bottom, letting this bug stay undiscovered for quite some time.